Add a dockerfile
Support docker builds with Ubuntu 24.04 image and Rust nightly
toolchain (2025-01-25 1.86.0). The dockerfile is installing toolchains
and other tools that are needed. VECTOR SDK rust-sdk-demo source code
and the SDK itself are also cloned to the docker image (path
/opt/src/rust-sdk-demo).

The dockerfile is installing a nightly toolchain 2025-01-25. It seems
that the latest nightly toolchain triggers an error. The latest version
that still works is 2025-03-18. The version 2025-03-19 generates an
error that was reported by hertelukas.

Signed-off-by: Markku Kylänpää <markku.kylanpaa@vtt.fi>